+++ libs/data-access-timeline/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/timeline.service';
 export * from './lib/threads.service';
 export * from './lib/url-preview.service';
+export * from './lib/edit-history.service';