+++ libs/util-matrix/src/index.ts
@@ -1,4 +1,6 @@
 export * from './lib/message-view';
+export * from './lib/edit-history';
+export * from './lib/edit-history-diff';
 export * from './lib/timeline-event';
 export * from './lib/media.model';
 export * from './lib/session.model';